backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
scikit-learn==1.5.0
numpy==1.26.4
reportlab==4.2.0
python-multipart==0.0.9
anthropic==0.29.0
python-dotenv==1.0.1
